[NFC] Reorganize ZHigh dialect: one cpp file for each operation (#1918)
* Each ZHigh op goes to a seperate file
