+++ src/Accelerators/NNPA/Dialect/ZHigh/CMakeLists.txt
@@ -1,18 +1,35 @@
+add_subdirectory(ZHighOps)
+
 add_onnx_mlir_dialect(ZHigh zhigh)
 add_onnx_mlir_dialect_doc(zhigh ZHigh.td)
 
-add_onnx_mlir_rewriter(ZHighCombine)
-
 add_onnx_mlir_library(OMZHighOps
-  ZHighHelper.cpp
+  # Top files for ZHigh dialect.
   ZHighOps.cpp
-  ZHighCombine.cpp
-  ZHighShapeHelper.cpp
+  ZHighOps/OpHelper.cpp
+  ZHighOps/ShapeHelper.cpp
+
+  # Support for shape inference, verifier, and canonicalization.
+  # Please list in the alphabetical order.
+  ZHighOps/BatchNorm/BatchNorm.cpp
+  ZHighOps/Conv2D/Conv2D.cpp
+  ZHighOps/Elementwise/Elementwise.cpp
+  ZHighOps/GRU/GRU.cpp
+  ZHighOps/LSTM/LSTM.cpp
+  ZHighOps/MatMul/MatMul.cpp
+  ZHighOps/MeanReduce2D/MeanReduce2D.cpp
+  ZHighOps/Pooling/Pooling.cpp
+  ZHighOps/Softmax/Softmax.cpp
+  ZHighOps/Stick/Stick.cpp
+  ZHighOps/StickForGRU/StickForGRU.cpp
+  ZHighOps/StickForLSTM/StickForLSTM.cpp
+  ZHighOps/Unstick/Unstick.cpp
 
   DEPENDS
   OMLayoutHelper
   OMZHighIncGen
-  OMONNXZHighCombineIncGen
+  OMONNXZHighStickIncGen
+  OMONNXZHighUnstickIncGen
   OMShapeInferenceOpInterfaceIncGen
   libzdnn
