+++ src/Accelerators/NNPA/Dialect/ZHigh/ZHighOps/ShapeHelper.cpp
@@ -2,7 +2,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-//===----------------ZHighShapeHelper.cpp - help for shapes ---------------===//
+//===----------------ShapeHelper.cpp - shape helpers for ZHigh ------------===//
 //
 // Copyright 2019-2022 The IBM Research Authors.
 //
@@ -12,7 +12,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "src/Accelerators/NNPA/Dialect/ZHigh/ZHighShapeHelper.hpp"
+#include "src/Accelerators/NNPA/Dialect/ZHigh//ZHighOps/ShapeHelper.hpp"
 #include "src/Accelerators/NNPA/Support/LayoutHelper.hpp"
 #include "src/Dialect/ONNX/ONNXOps/OpHelper.hpp"
 #include "src/Support/Diagnostic.hpp"
